src/server/requirements-dev.txt
pytest==8.4.2
pytest-flask==1.3.0
coverage==7.4.0
pytest-cov==4.1.0